+++ src/main/kotlin/com/bibireden/data_attributes/api/DataAttributesAPI.kt
@@ -1,7 +1,12 @@
 package com.bibireden.data_attributes.api
 
+import com.bibireden.data_attributes.DataAttributes
+import com.bibireden.data_attributes.config.functions.AttributeFunction
+import com.bibireden.data_attributes.config.models.OverridesConfigModel
+import com.bibireden.data_attributes.data.EntityTypeData
 import net.minecraft.entity.LivingEntity
 import net.minecraft.entity.attribute.EntityAttribute
+import net.minecraft.registry.Registries
 import java.util.*
 import java.util.function.Supplier
 
@@ -43,5 +48,20 @@ object DataAttributesAPI {
         else {
             Optional.empty()
         }
+    }
+
+    @JvmStatic
+    fun getOverride(attribute: EntityAttribute): OverridesConfigModel.AttributeOverride? {
+        return DataAttributes.OVERRIDES_CONFIG.overrides[Registries.ATTRIBUTE.getId(attribute)]
+    }
+
+    @JvmStatic
+    fun getFunctions(attribute: EntityAttribute): List<AttributeFunction>? {
+        return DataAttributes.FUNCTIONS_CONFIG.functions.data[Registries.ATTRIBUTE.getId(attribute)]
+    }
+
+    @JvmStatic
+    fun getEntityTypes(attribute: EntityAttribute): EntityTypeData? {
+        return DataAttributes.ENTITY_TYPES_CONFIG.entity_types[Registries.ATTRIBUTE.getId(attribute)]
     }
 }