+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+## Additions 💫
+- Added in the API a way to fetch functions, entity-types, and overrides from an `EntityAttribute`.
 ## Changes ⚙️
-- Addressed issue with the `DefaultAttributeFactory` continuing to attach extra config entries, causing unneeded duplications.
+- Resolved synchronization issues with the config manager, which caused issues when running DataAttributes on a server instance and connecting from a client.